Added handles to printer enclosure assemblies
I also fixed and organized some mates, removed the aluminum trim from the doors, and redid some naming conventions so names make more sense. To me.

I mean also I think to most people because the door configuration labeled "Right Top" is for the top right corner of a printer enclosure. It just means that the HANDLE for the top right door is then labeled "Right Bottom" because it's for the right side door but goes at the bottom corner of the panel. Which isn't really optimal and could possibly create some confusion, but the door handle top/bottom chirality indicators were created first and the handles have already been printed with the chirality indicators on them, and having the top door assemblies being labeled as bottom in CAD feels more wrong and also overall more confusing.

It's really not *that* bad and hopefully nobody needs to make major changes to this CAD design anyways, but if someone does need to look at or change things and gets confused that's why this note is here.