Designed new tall printer shelves for enclosed XL
There is now Tall Printer Frame V2a which drops the XL by only 2 pegs so it has exactly enough vertical clearance but removes the space for more i3-sized printers at the base of the shelves. There is also V2b which moves the shelf down by a lot more so that the far left shelf is able to house two i3-sized printers vertically in the same height that houses the XL in the center left shelf, which then means we don't reduce the total amount of i3-sized printers. V2b does reduced the amount of storage below the shelves more than V2a, but it is still tall enough for some space improvement storage and finished print bins.