Branding fidget cube progress
Created some test prints to try on the XL next time it's open. We probably should NOT start immediate production even if they print well because we should test with the 0.25mm nozzles first but like the files are here I guess.

Also the Lineart and Two Color versions of the logo are NOT final and were quickly thrown together just so I had something to try putting on the cube as a test. They are VERY BAD and some of the transparency layers are broken but I will make actual versions this or next weekend and these ones were good enough to throw on the cube temporarily.